Coset MDS implementation
This is almost like calling `Radix2Bowers`'s `coset_lde` with `added_bits=0`. I tried to reuse that code, but couldn't get comparable speeds. I haven't dug into the asm but I think there seem to be a couple issues:
- The compiler fails to do constant propagation, even in some cases which seem like they should be easy. This is why I explicitly precomputed `weights`. Ideally we'd compute them in `permute_mut` and constant propagation would make them immediate values, but it wasn't being done.
- I guess there are some range checks, and loops over matrix columns, that the compiler can't eliminate unless it inlines some large methods.

I think we may want the two implementations may diverge anyway, e.g. since `Radix2Bowers` is designed to be applied to matrices, we can vectorize over columns. `CosetMds` is for single polynomials, so we may want to vectorize like we did in Plonky2's FFT code.

Currently this is a lot faster than `apply_circulant_fft`, but I think in principle a similarly optimized convolution impl should be as fast, as the steps are essentially the same. With convolutions we have flexibility about weights (any circulant MDS row), but with `CosetLde` we also have flexibility from the coset choice, and from the fact that e.g. constant scaling doesn't change MDS.