Unified Bowers FFT
We now support both "G" and "G^T" FFTs from the Bowers paper. Instead of exposing both as separate options, we expose a single `Radix2Bowers`.  For single [i]DFTs, it doesn't really matter which we use.

For LDEs, we want to chain G with G^T, so that the two bit reversals cancel out.  This makes zero-padding slightly more awkward, we just have to do it while in bit-reversed form. Similarly in `coset_lde_batch`, we want to rescale coefficients by shift powers while staying in bit-reversed form.

Gives around a 10% speedup for coset FFTs.