`exp_const_u64` (#79)
* `exp_const_u64`

This doesn't necessarily feel like the right solution, but in my builds it does seem to speed up Poseidon benchmarks significantly.

OTOH Hamish earlier found that the compiler seemed to be specializing the `exp_u64` code for constant arguments as one would hope. So we should probably look into this more. Not sure if it's a compiler difference (I'm using a recent nightly and targeting M1) or due to differences in the call sites.

* Change cube default