`AbstractField` changes (#122)
The main change is to merge `AbstractField` and `AbstractionOf`, simplifying the trait hierarchy.

Previously `AbstractField` didn't know what concrete field it was an abstraction of; now it does via an associated type `F`. This lets us write code like this in `AbstractField`
```rust
fn exp_u64(&self, power: u64) -> Self {
    Self::F::exp_u64_generic(self.clone(), power)
}
```
which was the original motivation for this change. It's a bit of a strange API, but it lets us override `exp_u64_generic` with addition chains that are relevant to a particular field, and have those optimized addition chains used for all abstractions of that field, such as `FieldArray`.