Improved Karatsuba Decomposition
The Even-Odd decomposition seems to be faster than the Left-Right decomposition.

Additionally extended the code to work with matrices of size 64x64.

Still currently quite a bit of copy and pasted code so it would be good to find a way to cut down on that. Also seems to be doing a bunch of messing around with data. (E.g. similar to bit reversals which turn up in FFT's). Ideally we would find a way to avoid these.