Code Refactoring
Continued Large refactoring of Code.

Introduced 2 new types, corresponding to the trait NonCanonicalPrimeField32 and started working on making the convolution code more polymorphic which should let us reduce how much code is currently there. (In particular we should be able to get rid of all the large_entries functions.)

More work to do but the fundamentals are coming together.

I'm also wondering if NonCanonicalPrimeField32 might be able to be used in FFT's as well to help implement delayed reductions. (But that it for a future project)