Finished Convolution Updates
It should be now possibly to switch from convolutions in M31 to convolutions in BB essentially for free. Speed seems to be unaffected.

Was also able to simplify code to eliminate some repartitions.

Still need to update code for non power of 2 sizes but this should be an easy change.