Fix the domain we evaluate the quotient over (#198)
Previously we were evaluating the quotient polynomial over the same domain that trace LDEs were defined over. However, it's valid and sometimes useful to let them diverge, i.e. with a larger blowup factor to minimize FRI queries.

It's also occasionally useful to have the opposite - a quotient whose degree exceeds our (committed) blowup factor. In that case we'd do the larger LDE for quotient evaluation, but Merklize a strided view of it, to speed up the commitment. However, this is a bit harder to support, and probably less frequently useful.