Adding Structure
Adding file structures. Currently a bunch of empty files but will over time move the code in lib to the relevant files. Goal is to make structure identical to the dft module.

Added a trait into Field to characterize ComplexExtensions. Not sure if this is the right approach but trying to make something somewhat analogous to TwoAdicField.