Circle Fast Fourier Transform (#209)
* Circle Fast Fourier Transform

First push for setting up code to compute the Circle Fast Fourier Transform over M31.

Currently the code works and is fast but need to be updated to better match the rest of the API.

* Removing Duplicate Code

Powers is already defined in field so we just import it from there.

* Adding Structure

Adding file structures. Currently a bunch of empty files but will over time move the code in lib to the relevant files. Goal is to make structure identical to the dft module.

Added a trait into Field to characterize ComplexExtensions. Not sure if this is the right approach but trying to make something somewhat analogous to TwoAdicField.

* Fleshing out the ComplexExtension Trait

This allows us to remove the use of both Point and TwoAdicField in the cfft.

Also some minor fmt/clippy fixes.

* Moving more functions out of lib

Beginning the process of rewriting CFFT to use RowMajorMatrix instead of vec.

Currently completely broken.

* Radix 2 CFFT is now functional

Updated Radix 2 CFFT to work with Row Major Matrix and it all seems to work now at least if the input is a vector.

Currently only testing vector entries but will update this testing suite once the Inverse FFT is also working.

* Got Inverse Cfft working

Just coset cfft remains but that should be easy.

Also think I've fixed the issue which caused the build error. Required adding a new file "old" but this will be removed before we merge.

* Fixing formatting

The formatter is unhappy about the ordering of imports for some reason.

* fmt

* coset extrapolation and testing suite

Added Coset extrapolation and improved the testing suite.

* Improved Tests

* fmt

* fmt

* Final changes

Set up testing for coset_lde and got the trait working.

Removed LDE as it doesn't really make sense in this setting.

Removed icfft_batch as the algorithm is essentially identical to coset_icfft_batch.

Adding some comments.

* Minor changes thanks to review