Delayed Reduction for BabyBear Multiplication
It is possible to speed up 3 and 4 way multiplications in BabyBear using delayed reduction.

In particular this can be used to speed up the `x -> x^7` map by roughly `30%`.

As this map is used in Poseidon2 it gives roughly a `10%` speed up for the Poseidon2 hash.

It should also apply to the rescue hash but that will require rewriting the current way that the rescue hash sbox layer works to use exp_const_u64 not exp_u64.