Delayed Reduction for 2-Exponentiation
When computing x^{2^n} as a chain of squares, we don't need to fully reduce each time. It's good enough to reduce to a value in (-P, P) as after squaring we always get something positive anyway.

This gives a ~17% speed up for try_inverse and a ~25% speed up for root_7