Minor Refactor
Added a constant method in goldilocks to convert a constant array into a constant array of goldilocks elements.

This makes the: matmul_internal method again identical between the two fields so we can store it in the Poseidon2 crate for now. Long term we likely want to specialize this further.