Karatsuba-based convolution functions for MDS matrix multiplication (#248)
* Benchmarking for FieldArray Exp

as title

* Improving Sub and Add in M31

After chatting with Jacqui I thought I'd try and make a couple of minor improvements to the base add/sub methods.

Sub seems to be ~20% faster, add is closer to 5% faster.

* Avoiding a proliferation of bench files.

* Update Cargo.toml

* Re-adding bench for add/sub

* Combining Bench File

Made a new benchmark file to combine the benches for add, sub, inverse and 5th root.

Fixed some formatting issues.

Also replaced overflowing_add_unsigned by overflowing_add as it seems to be faster for some reason.

* clippy

* Cleaning Up

Removing the old commented out implementation of sum and replacing it with a link.

* Further Optimisations

Some black magic optimisations thanks to Jacqui.

Improves the compiled assembly output.

* Small MDS 16x16 matrix

Initial commit, many fixes left to do

* Added General Functions

Adding general functions to compute the convolution using a Karatsuba approach for all 2^n. Currently these are much much slower than the specialized functions (20x or more slower). Need to keep investigating why this is the case.

* Improved Karatsuba Algorithms

Improved the Karatsuba algorithms to perform a single step of the FFT algorithm before resorting to Karatsuba methods.

This improves constants by a reasonable margin and get us to the point where the Karatsuba 16 convolution is now 25-30% faster than the n^2 implementation.

Additionally implemented this method for convolutions of size 32 making use of i128's. This gives a massive (> 50%) speed up over the current implementation as it allows for delayed reductions.

* Added: 'from_wrapped_u128' to field specification.

Naively extended all the from_wrapped_u64 methods to from_wrapped_u128. These methods could likely be improved.

* Added Comments

Added a lot of comments to all functions.

Introduced a new type SimpleInteger so that the functions in here will work with both i64 and i128 as well as potentially other types.

Cleaned up some of the code and moved things around somewhat.

* Improved Karatsuba Decomposition

The Even-Odd decomposition seems to be faster than the Left-Right decomposition.

Additionally extended the code to work with matrices of size 64x64.

Still currently quite a bit of copy and pasted code so it would be good to find a way to cut down on that. Also seems to be doing a bunch of messing around with data. (E.g. similar to bit reversals which turn up in FFT's). Ideally we would find a way to avoid these.

* Formatting

* Minor Updates

Moved code around somewhat to make it easier to follow.

Slightly improved the recombination code for convolutions to use less additions and use less variables by passing in a mutable output vector.

* Formatting

cargo fmt...

* Updated Large Convolutions

Added code to speed up large convolutions, allowing us to work basically just with i64's.

Unfortunately this involved duplicating a bunch of functions. Will try to clean it up over the next couple of days.

* Improving reductions

Added specific code to do the reductions in the final step.

Minor formatting issues.

* Minor fixes

Cleaned up the code which converted back to a Field element in the last step.

Added comments and debug_assert.

* Setting up code for babybear

Starting work on a major overhaul to allow code to work with multiple fields.

* Completeing Proof

Completed the proof that the suggested algorithm is valid.

* Code Refactoring

Continued Large refactoring of Code.

Introduced 2 new types, corresponding to the trait NonCanonicalPrimeField32 and started working on making the convolution code more polymorphic which should let us reduce how much code is currently there. (In particular we should be able to get rid of all the large_entries functions.)

More work to do but the fundamentals are coming together.

I'm also wondering if NonCanonicalPrimeField32 might be able to be used in FFT's as well to help implement delayed reductions. (But that it for a future project)

* Finished Convolution Updates

It should be now possibly to switch from convolutions in M31 to convolutions in BB essentially for free. Speed seems to be unaffected.

Was also able to simplify code to eliminate some repartitions.

Still need to update code for non power of 2 sizes but this should be an easy change.

* Ported BabyBear convolutions over

Massive speed up on the current implementations, particularly for 32 and 64 cases.

Also got the n = 12 case done.

* Formatting

* Added some dot products.

* Initial Attempt at Noncanonical FFTS

Initial try to see if we can easily implement delayed reduction for FFT's.

Had to shuffle around code to prevent a circular dependancy chain.

* More Generic Functions

Implementing Hamish's idea to reduce the amount of code needed for Karatsuba Convolutions.

* Passing in Output

Minor change to have the output passed in. More tests to come on this.

Also fixed some formatting and deleted some now unneeded code.

* Adding debug checks and unsafe markers

As title

* Minor formatting and comments

* Rolling back DFT changes.

Removing the DFT changes. Might do something in a future pull request but not now.

* Formatting

Forgot to run Cargo fmt on the last upload... Also removing code I added to dft/naive.

* Removing changes to dft/traits.

This should be the last file I need to roll back.

* Use faster reduction in Goldilocks.

* Move field-specific MDS code to the fields' crates.

* Move field-specific MDS code to the respective field trait.

* Finish moving MDS code to field traits; miscellaneous tidying.

* Update bench imports after move.

* Comments and fiddling with memory loads.

* Faster/cleaner i64 and i128 reductions.

* Better reduction; implement size 24; more benchmarks.

* Refactor `Convolve::apply()`.

* Use small size 16 MDS matrix for Goldilocks.

* Update expected output for Goldilocks MDS 16 after changing matrix.

* Simplify reduction since the final values must be non-negative.

* Expand documentation.

* Update Rescue Prime expected outputs to use new size 12 MDS matrix.

* Cargo fmt

* Remove old non-canonical files.

* Remove left-over debugging info.

* Remove now unused linear_combination methods.

* Replace `mul` with `parity_dot` as main template method.

* Use Angus's faster partial reduction trick.

* Add `parity_dot` for BabyBear and associated reduction.

* Minor tidying.

* Encourage rather than force inlining for large convlutions.

* Add two functions missed in the merge.

* Clippy.

* Documentation.

* Implement `from_wrapped_u128` for NEON `PackedField`s

* Tidy comments following Angus's suggestions.

* Remove unused `from_wrapped_u128`.

* Update mds/src/karatsuba_convolution.rs

Co-authored-by: Jacqueline Nabaglo <jnabaglo@polygon.technology>

* Replacing Signed by Negacyclic

Signed Convolutions already have a name, Negacyclic!

This update is simply an in place replacement of every instance of signed by negacyclic.

* Minor Fix

This should make all the tests pass.

* Add debug assert and comment

The dot_product function will fail if N = 0 but this would be a little annoying to fix and should never come up.

* Make conv_n, negacyclic_conv_n private

Moving conv_n and negacyclic_conv_n out of the trait to make the functions private.

---------

Co-authored-by: AngusG <angusgruen@gmail.com>
Co-authored-by: Jakub Nabaglo <jakub@mirprotocol.org>
Co-authored-by: Daniel Lubarov <daniel@lubarov.com>
Co-authored-by: AngusG <agruen@polygon.technology>
Co-authored-by: Jacqueline Nabaglo <jnabaglo@polygon.technology>