Poseidon2 speedups (#257)
* Adding Sum Benchmark

Adding a benchmark to see how fast the current iterator summation functions are.

(Suspect this might be one of the bottlenecks)

* Code to test Summations

Simple testing code to look at delayed reduction in summations. As can be seen, delayed reduction is much faster for larger sizes (16 and up).

Thus we should be using this for larger sizes for Poseidon2.

* Removing MDS dependency from Poseidon 2

We never used MDS but needed features = ["min_const_gen"] in rand which was being gotten indirectly from loading in the MDS crate.

* Refactor Babybear

Moving all BabyBear specifics into the BabyBear crate.

This will allow us to access the internals without having to call to_monty or from_monty.

* First Algorithm Updates

Added delayed reduction for the summation and saved the multiplicative constants already in MONTY form.

~20% speed up for WIDTH 16
~25% speed up for WIDTH 24.

* Testing faster exp^7 methods

Seems like I can get ~ a 10% improvement. Possibly more can be squeezed out with some refinements to monty_square_reduce and monty_cube_reduce.

The numbers in the multi_mul benchmarks seem to be quite misleading and disagree a lot with the numbers in the poseidon2 benchmark.

* Speed Improvments

Slightly improved the monty_square_reduce algorithm. Using this to compute `x -> x^7` we get a `~25 - 30%` speed improvement over the current implementation.

This leads to an immediate `10-15%` extra speed up for Poseidon2 and should also speed up any other algorithms using the `x -> x^7` map.

It can also probably be used to improve the inverse and root_7 functions though this will be more work.

* Fixing bench

Making the Exp 7 benchmark correctly compare the three different natural methods.

* Fixing Constants

Added a constant function to convert an array of u32's into an array of BabyBear elements in Monty form.

Needed as the ToMonty function can changes on different architectures.

* Removing delayed reduction code

Going to deal with that in a different pull request.

* Cleaning up Sum Benchmarking Code

I've been informed that benchmarks in the single ns are a bit BS so added a loop to push all timings into the microseconds.

* Refactored Goldilocks

Moved the Goldilocks specific code into the Goldilocks module.

* fmt and added comment

* Fixing Imports

* More small fixes

Hope these are the last. Just making sure all imports of DiffusionMatrixBabybear come from the right place.

* Using delayed summation for Goldilocks

Slightly faster for sizes >= 16.

Also fixed Clippy Error.

* Cargo fmt....

* Update baby-bear/src/baby_bear.rs

Co-authored-by: Hamish Ivey-Law <426294+unzvfu@users.noreply.github.com>

* Update baby-bear/src/baby_bear.rs

Co-authored-by: Hamish Ivey-Law <426294+unzvfu@users.noreply.github.com>

* Update baby-bear/src/baby_bear.rs

Co-authored-by: Hamish Ivey-Law <426294+unzvfu@users.noreply.github.com>

* Update poseidon2/src/matrix.rs

Co-authored-by: Hamish Ivey-Law <426294+unzvfu@users.noreply.github.com>

* Update poseidon2/src/matrix.rs

Co-authored-by: Hamish Ivey-Law <426294+unzvfu@users.noreply.github.com>

* Resolving most requested changes.

Changed sum benchmarks to take in an input giving number of repetitions and length of the summation.

Added comments above sum_u64/sum_u128 specifying when they should be used. (E.g. for what summation lengths they are faster.). Updated sum_u128 to use reduce128 instead of a %.

Fixed comment in baby-bear/poseidon2.

Added

rand = { version = "0.8.5", features = ["min_const_gen"] }

to [dev-dependencies] for babybear and goldilocks.

Technically these aren't needed as both field crates import rand and poseidon2 and poseidon2 imports min_const_gen so min_const_gen  ends up getting enabled in the field crates anyway. But this prevent strange errors from appearing if we at some point rework poseidon2 and remove min_const_gen from its imports.

* Fixing the interaction with Packed Fields

Accidentally broke things when trying to specialize to BabyBear.

* Fixing code to work with Packed Fields

This should fix the issue with non compiling when AVX/Neon is enabled.

To fix this I had to roll back the summation changes I made, I'll push them in a separate PR to do all three fields at once.

* Minor Refactor

Added a constant method in goldilocks to convert a constant array into a constant array of goldilocks elements.

This makes the: matmul_internal method again identical between the two fields so we can store it in the Poseidon2 crate for now. Long term we likely want to specialize this further.

* Minor Typo Fixes

---------

Co-authored-by: Hamish Ivey-Law <426294+unzvfu@users.noreply.github.com>
Co-authored-by: Hamish Ivey-Law <hamish@ivey-law.name>