Initial Commit for supporting a new Field.
For lots of arithmetic hashes, it's useful if x -> x^3 is a permutation of the multiplicative group (Note this isn't the case for M31 or BB).

The KoalaBear prime is 2^31 - 2^24 + 1 = 127 * 2^24 + 1 which means the cubic map will be a permutation.

Outside of allowing for the cubic map, the code/performance should be essentially identical to that of BabyBear (Hence the similar name). The main downside is that it has slightly lower 2-adicity (24 as opposed to 27).

Currently missing a lot of code but this will be updated over the next few days.