Cleaning up tests and round constants (#299)
* Cleaning up tests

Removing the import of HorizenLabs Poseidon2 implementation from BabyBear and Goldilocks tests.

It still gets imported in the bn254 tests as that version of Poseidon2 is a little different (Width is not a multiple of 4) so I don't currently have a sage implementation of it. Not sure how mush we are planning to support bn254 anyway so just going to leave that as is for now.

Additionally, inside Poseidon2, split apart the constants into external and internal round constants. The external constants are each the form [F; WIDTH] where as the internal ones are directly elements of F so it doesn't really make sense to save the constants together.

* fmt fix

* Update round_constants.rs

Typo

* Update poseidon2/src/round_constants.rs

Co-authored-by: Hamish Ivey-Law <426294+unzvfu@users.noreply.github.com>

* Update bn254-fr/src/poseidon2.rs

Co-authored-by: Hamish Ivey-Law <426294+unzvfu@users.noreply.github.com>

* Update lib.rs

* Update round_constants.rs

* Minor fixes

Minor fixes to get constant generation to work.

* Moving round_constants into respective field crates.

* Update poseidon2.rs

Removing unneeded comments

---------

Co-authored-by: Hamish Ivey-Law <426294+unzvfu@users.noreply.github.com>