Added trait, MDSLightPermuation
This trait is the property that a matrix needs to satisfy in order to be used as the external_linear_layer in the external (full) rounds of Poseidon2.

Currently this trait is not used but in a future PR we will modify Poseidon2 to no longer hardcode  the external_linear_layers.

This will allow us to optimise this matrix both generally and for specific fields/architectures.