Better External Linear Layer
We can improve the External Linear layer to require only 9 additions and 2 doubles as opposed to 10 additions and 4 doubles.

Only a small speed up (Roughly 2-5%) but no point leaving simple improvements on the table.

Currently still supporting both options to give the option to match our implementation with the Horizon Labs implementation.