Algorithm Improvements
This implements delayed reduction for Poseidon2 M31. The compiler also now seems to be able to unroll this loop.

Seems to be ~20% faster. (About 5-10% of that coming from the #[inline])

Also fixing MATRIX_DIAG_16_MONTY_SHIFTS, there is no need for this to be i32. Leaving it as u64 so it can be used with MATRIX_DIAG_16_MONTY_SHIFTS if wanted.