KoalaBear, a possible alternative to BabyBear (#329)
* Initial Commit for supporting a new Field.

For lots of arithmetic hashes, it's useful if x -> x^3 is a permutation of the multiplicative group (Note this isn't the case for M31 or BB).

The KoalaBear prime is 2^31 - 2^24 + 1 = 127 * 2^24 + 1 which means the cubic map will be a permutation.

Outside of allowing for the cubic map, the code/performance should be essentially identical to that of BabyBear (Hence the similar name). The main downside is that it has slightly lower 2-adicity (24 as opposed to 27).

Currently missing a lot of code but this will be updated over the next few days.

* Added Two Adicity Information

* Adding  description of Primes

Adding info regarding why BabyBear/KoalaBear are the specific values given.

Also adding some notes for alternatives to KoalaBear which might be useful in different cases.

* Added Extension Fields

Added code to support binomial extensions of KoalaBear.

Unfortunately no Binomial extension of degree 5 exists so currently just defining 2 and 4.

If we need an extension of degree 5 might have to include that manually.

* Godbolt Checks

Double checked that all the base functions are compiling to what we would expect. (i.e. the same as the BabyBear code with slightly changed constants).

* Adding support for Poseidon2

Going to make some other changes to Poseidon2 separately.

* AVX and NEON implementations

* Fix multiplicative inverse tests

* Poseidon2 Implementation

Also some file updates to match changes in main.

* fmt

* Adding Poseidon2 for Packings

* Update readability

* typo

* Update To newer version

Some extra work had been done in the Neon BabyBear file which can/should be reflected in the koala-bear file.

* fmt

* typo fixes

---------

Co-authored-by: Jacqueline Nabaglo <jnabaglo@polygon.technology>