Observe the final polynomial in FRI (#393)
There are some challenger related complications here, since the `SerializingChallenger`s don't impl `CanObserve` for extension fields, and doing so creates a theoretical conflict with `Hash`.

The workaround I'm proposing here is to bring back `FieldChallenger` and use its extension field methods. The downside is that the core FRI code now needs to know about the base field `Val`, so we can work with `FieldChallenger<Val>`.