Adding custom AVX2 implementations for x^3, x^5 and x^7 for Monty31Fields (#437)
* Adding custom AVX2 implementations for x^3, x^5 and x^7 for Monty31Fields.

This is part of the work to speed up Poseidon2 which can be done immediately without needing any API changes. At some point I'll implement this for AVX512/NEON but there is no need to wait for that in order to merge this.

Looking at GodBolt: https://godbolt.org/z/Ksx1z1Whe we see that in pure number of instructions, we get a little less than a 10% improvement for x^3 and this climbs to a 20% improvement for x^7.

As x^3 and x^7 are the sbox used by KoalaBear and BabyBear respectively for their Poseidon2 hash, this should give a mild improvement to that.

Indeed we find that we get close to a 10% speed up for Poseidon2 BabyBear but broadly an unnoticeable speed up for Poseidon2 KoalaBear. This is likely due to the x -> x^7 being a major bottle neck for Poseidon2 BabyBear whereas Poseidon2 KoalaBear is currently being slowed down more by the current (poor) implementation of the internal layer with the map x -> x^3 already being reasonably quick.

* Moving all testing of exp 3, 5, 7 into packed_field_testing.

Also allows us to delete a couple of testing functions in the NEON code.

* Update monty-31/src/x86_64_avx2/packing.rs

Co-authored-by: Hamish Ivey-Law <426294+unzvfu@users.noreply.github.com>

* Update monty-31/src/x86_64_avx2/packing.rs

Co-authored-by: Hamish Ivey-Law <426294+unzvfu@users.noreply.github.com>

* Update monty-31/src/x86_64_avx2/packing.rs

Co-authored-by: Hamish Ivey-Law <426294+unzvfu@users.noreply.github.com>

* Update monty-31/src/x86_64_avx2/packing.rs

Co-authored-by: Hamish Ivey-Law <426294+unzvfu@users.noreply.github.com>

* Minor fixes to align to review.

* Minor comment updates

---------

Co-authored-by: Hamish Ivey-Law <426294+unzvfu@users.noreply.github.com>