Parallelize batch inversion
This is sort of naive as it involves several separate instances of the batch inversion algorithm, resulting in several `inverse()` calls. It would be more optimal to have one big parallelized instance, with one big partial product tree made up of smaller per-thread trees, but that seems a bit annoying to code. Inversion isn't terribly expensive (especially for 32-bit fields) so the difference doesn't seem that important, but it could be worth revisiting at some point.

Separately, this should ideally be vectorized. I tried doing so with `F::Packing`, but it didn't help much as the most important callers use extension fields with trivial packings. We could use the `ExtensionPacking` approach, there are just some API complications.

This change alone reduces inversion costs from ~6% to ~1%, so pretty good for now.