Generalize the basic Monty trick impl
And have the "chunked" implementation (the one with better instruction-level parallelism) call it with a `FieldArray`.

For now, I also changed `FieldArray` to implement `PackedValue`, but I'm not sure about it. There might not be that much benefit (just slightly convenient to have `pack_slice`), and it does fail to enforce `PackedValue`'s power-of-two requirement, though I can't remember why that's needed.