AVX512 implementation for Koala/BabyBear, Poseidon2 (#498)
* First attempt at updating the MONTY31 code for AVX512.

Need to find a way to map across signed_add_avx512 efficiently. Unfortunately AVX512 not having _sign_ is slightly painful.

* Typo's

* Adding a correct but poor solution for now.

I suspect there should be a way to reduce the number of instructions used in signed_add by 1.

* Code sorta working but Godbolt suggests the compiled AVX512 is a little strange...

https://godbolt.org/z/Gsr868bWT

* Poseidon2 AVX512 implementation for Monty31 Fields.

Still a little work to be done as there are some strange compilation issues godbolt is pointing to. But the code is working and relatively fast now.

(At the very least it's much faster than 2*AVX2 code speed which we definately expect).

* Saving an instruction.

Code is looking good but comments need alot of work.

* Updates to comments.

* Formatting

* Fixing godbolt link.

Also removed a couple of uneeded godbolt links in babybear and koalabear.

* Adding comments and minor bug fixes.

* Update monty-31/src/x86_64_avx2/utils.rs

Co-authored-by: Hamish Ivey-Law <426294+unzvfu@users.noreply.github.com>

* Comment fixes +avoiding a couple of clones.

---------

Co-authored-by: Hamish Ivey-Law <426294+unzvfu@users.noreply.github.com>