Modify `Radix2DitParallel` to treat a coset LDE... (#527)
* Modify `Radix2DitParallel` to treat a coset LDE...

... as the union of several DFTs onto several smaller cosets.

Also change it to resize the input vector rather than allocating a separate vector for the result. This works well iff the input vector was already allocated with extra capacity.

Also do the first layer of most coset DFTs out-of-place. This is effectively like merging the steps where we copy coefficients onto other cosets into the next layers that use those coefficients.

* Alternate forwards and backwards

* Fix

* Refactor

* Merge division by height