Bn254 Binary Euclidean Inversion (#920)
* Initial work based on a Paper.

Unfortunately ran into a couple of minor issues. Might need to try a slightly different approach as BN254 doesn't have cheap reductions.

* Super simple GCD method.

This can be improved a lot but is a good proof oc concept which compiles and passes tests.

Gives a ~10% speed up over the current method.

* delete old method as it's worse even before optimisations...

* Saving a mul

* Minor changes though these don't really help rn.

* Slow working towards faster but more complex gcd. Need to update from main.

* GCD implementation now working!

* Cleaning up. First stages.

* Comments and some minor simplifications

* Comment

* rm uneeded changes

* typo fix

* rm halve code.

Moved that to a different PR

* Adding comments,

refactor a couple of functions to reduce the number of lines of code and remove if statements.

* oops, left my comment out of #![no_std] in.

* Lots more and better comments

* Improving comments more