Implement uniform sampling of bits from field elements (#1050)
* implement `UniformSamplingField` and `CanSampleUniformBits` traits

The former is the trait a field will implement to define the
parameters needed for the `sample_uniform_bits` function via the
`CanSampleUniformBits` (which the challengers need to implement).

* implement `UniformSamplingField` for all small fields

* implement `CanSampleUniformBits` for `DuplexChallenger`

This also adds the `uniform-sampling-may-panic` feature, which decides
whether finding a field element >= m_k will resample (rejection
sampling) or panic. In some contexts, e.g. recursive zkVMs it may be
preferable to panic and just accept a very small probability of a
proof failing.

* implement `UniformGrindingChallenger`

This goes hand in hand with the uniform sampling for the duplex
challenger. Instead of grinding using regular sampling, this one
samples using uniform sampling.

* fix regression in `sample_bits`

Accidentally cut off the normal implementation, oops

* split sampling of values into may panic / resampling approach

We'll make it so that the user can choose which variant to call.

* remove feature for sampling with panic, split into two functions

`_may_panic` suffix for functions that may trigger a panic.

* implement `_may_panic` variants for UniformGrindingChallenger

As it is now not a feature flag, but requires a different function
call, we simply add different overloads for variants that may panic to
make it clear at the callsite.

* implement UniformSamplingField for Monty31 field

Due to the orphan rule we have to implement this trait either in the
challenger trait or in the Monty31 crate.

It seems to make sense to implement this in the challenger, as the
Monty crate is more fundamental than the challenger, which anyway
requires the notion of fields.

* use `Self` for BabyBear / KoalaBear prime

* remove outdated grinding challenger code

* remove duplicated doc comments for UniformSamplingField for each field

* remove doc comment for blanket implementation

We keep a basic comment to clarify the reader as to why it exists.

* implement sampling of bits based on a strategy-like pattern

* implement uniform grinding with generic fn

* check bits are >0 in grind fns

* add doc comments to `check_witness[_may_panic]`

* fix sort order of dependencies

* apply fixes for `cargo +stable clippy --all-targets -- -D -warnings`

* address cargo +nightly fmt --all fixes

* improve doc comment about m_k

* allow `bits = 0` argument, just return zero always

* remove `_strategy` argument, use explicit generic call instead

* have single `sample_value` function with a PANIC_ON_REJECTION const

* switch to Result based API for uniform sampling of bits

In order to allow the verifier called functions to just return false
if a panic would have happened on the prover's side

* have single `sample_uniform_bits` with const generic to choose

* allow bits == 0 in grind, grind_uniform again

* add doc comment about CanSampleUniformBits trait

* rename all panic -> error

* remove bits > 0 from last grind fn as well

* address Clippy errors

* cargo fmt fixes